refactor: move map.jsx to src/
